refactor(benchmarks): extract benchmark types to benchmarks/_benchmark_types.py
Moved shared benchmark dataclasses, typed dictionaries, and validation logic
from benchmarks/pipeline_throughput.py into a new module benchmarks/_benchmark_types.py.
This improves code organization and separation of concerns.
